requirements.txt
fastapi
uvicorn
python-dotenv
Google-generativeai